backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
redis==5.0.1
aioredis==2.0.1
qrcode[pil]==7.4.2
pillow==10.1.0
httpx==0.25.2
python-dateutil==2.8.2
email-validator==2.1.0
jinja2==3.1.2
aiosmtplib==3.0.1
phonenumbers==8.13.26
pytz==2023.3
pandas==2.1.3
openpyxl==3.1.2
